tests/wpt/web-platform-tests/web-bundle/subresource-loading/script-relative-url-in-web-bundle.https.tentative.html

<!DOCTYPE html>
<title>Relative Url in web bundle</title>
<link rel="help" href="https://github.com/WICG/webpackage/blob/main/explainers/subresource-loading.md" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.js"></script>

<body>
  <script type="webbundle">
    {
      "source": "../resources/wbn/relative-url.wbn",
      "resources": [
        "relative-url-file.js",
        "relative-url/start-with-double-slash.js",
        "relative-url/start-with-slash.js",
        "relative-url/subdirectory-path.js",
        "starts-with-two-dots.js",
        "starts-with-two-dots-out-of-scope.js"
      ]
    }
  </script>
  <script>
    setup(() => {
      assert_true(HTMLScriptElement.supports('webbundle'));
    });

    const loaded_scripts = [];
    function scriptLoaded(file) {
      loaded_scripts.push(file);
    }
    const failed_scripts = [];
    function scriptFailed(file) {
      failed_scripts.push(file);
    }
  </script>
  <script src="../resources/wbn/relative-url-file.js"></script>
  <script src="../resources/wbn/relative-url/start-with-double-slash.js"></script>
  <script src="../resources/wbn/relative-url/start-with-slash.js"></script>
  <script src="../resources/wbn/relative-url/subdirectory-path.js"></script>
  <script src="../resources/wbn/starts-with-two-dots.js"></script>
  <script src="../resources/starts-with-two-dots-out-of-scope.js"
  onerror="scriptFailed('starts-with-two-dots-out-of-scope.js')"></script>

  <script src="resources/relative-url-in-web-bundle-test.js"></script>
</body>
